% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functionsa.R
\name{log_lkd}
\alias{log_lkd}
\title{Mixture likelihood function}
\usage{
log_lkd(y, w, mu, sigma)
}
\arguments{
\item{y}{vector of observations.}

\item{w}{vector of weights; they have to sum to one.}

\item{mu}{vector of location parameters for the Gaussian components.}

\item{sigma}{vector of standard deviation parameters for the Gaussian components.}
}
\value{
a scalar representing the log-likelihood of the mixture of Gaussian components.
}
\description{
This function compute the log-likelihood of a mixture model for
a fixed number of Gaussian components.
}
\keyword{BayesMIC}
