% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functionsa.R
\name{log_prior_pois}
\alias{log_prior_pois}
\title{Log-prior density, with a Poisson prior distribution representing the loss-based prior.}
\usage{
log_prior_pois(
  w,
  mu,
  sigma,
  rt_pois = 1,
  sh = 1.5,
  rt = 0.5,
  alpha_dir = 1,
  me_mu = 20,
  sd_mu = 10
)
}
\arguments{
\item{w}{vector of weights; they have to sum to one.}

\item{mu}{vector of location parameters for the Gaussian components.}

\item{sigma}{vector of standard deviation parameters for the Gaussian components.}

\item{rt_pois}{rate parameter of the Poisson prior distribution for the number of Gaussian components. Default value at 1.}

\item{sh}{shape parameter of the inverse gamma prior distribution for the standard deviation parameters of the Gaussian components. Default value at 1.5.}

\item{rt}{rate parameter of the inverse gamma prior distribution for the standard deviation parameters of the Gaussian components. Default value at 0.5.}

\item{alpha_dir}{parameter of the Dirichlet prior distribution for the weight parameters of the Gaussian components. Default value at 1.}

\item{me_mu}{mean parameter of the Gaussian prior distribution for the mean parameters of the Gaussian components. Default value at 20.}

\item{sd_mu}{standard deviation parameter of the Gaussian prior distribution for the standard deviation parameters of the Gaussian components. Default value at 10.}
}
\value{
a scalar representing the log-joint prior density of the parameters of a mixture of Gaussian components.
}
\description{
This function compute the log-joint prior for the parameters of a mixture model with an
unknown number of Gaussian components.
}
\keyword{BayesMIC}
