% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functionsa.R
\name{post_bnb}
\alias{post_bnb}
\title{Simulations from the joint posterior distribution of the parameters of a
mixture of Gaussian components with unknown number of components, by using a beta-negative-binomial prior for k.}
\usage{
post_bnb(
  y,
  K = 10^5,
  alpha_bnb = 1,
  beta_bnb = 1,
  sh = 1.5,
  rt = 0.5,
  alpha_dir = 1,
  me_mu = 20,
  sd_mu = 10
)
}
\arguments{
\item{y}{vector of observations.}

\item{K}{number of simulations. Default value at 10^5.}

\item{alpha_bnb}{shape parameter of the beta-negative-binomial. Default value at 1.}

\item{beta_bnb}{rate parameter of the beta-negative-binomial. Default value at 1.}

\item{sh}{shape parameter of the inverse gamma prior distribution for the standard deviation parameters of the Gaussian components. Default value at 1.5.}

\item{rt}{rate parameter of the inverse gamma prior distribution for the standard deviation parameters of the Gaussian components. Default value at 0.5.}

\item{alpha_dir}{parameter of the Dirichlet prior distribution for the weight parameters of the Gaussian components. Default value at 1.}

\item{me_mu}{mean parameter of the Gaussian prior distribution for the mean parameters of the Gaussian components. Default value at 20.}

\item{sd_mu}{standard deviation parameter of the Gaussian prior distribution for the standard deviation parameters of the Gaussian components. Default value at 10.}
}
\value{
k, a sample from the posterior distribution of the number of components.

param, a matrix from the posterior distribution of the weights, the means and the standard deviations of the components.

olp, a vector a joint prior density of the posterior samples.

oll, a vector a likelihood values of the posterior samples.
}
\description{
Simulations from the joint posterior distribution of the parameters of a
mixture of Gaussian components with unknown number of components, by using a beta-negative-binomial prior for k.
}
\keyword{BayesMIC}
